Simplified react ui structure
