+++ extension/ts/main-interface.js/App.scss
@@ -1,15 +1,3 @@
-.Demo {
-  display: flex;
-
-  & .Text {
-    flex: 1;
-    border: 1px solid var(--color-neutral-5);
-    margin-left: 120px;
-    padding: 40px 40px;
-    border-radius: 2px;
-  }
-}
-
 .BergamotApp {
   width: 100%;
   background-color: var(--color-neutral-2);
