+++ extension/ts/main-interface.js/components/Progress/_Progress.scss
@@ -1,3 +1,15 @@
+.DemoControlledProgress {
+  display: flex;
+
+  & .Text {
+    flex: 1;
+    border: 1px solid var(--color-neutral-5);
+    margin-left: 120px;
+    padding: 40px 40px;
+    border-radius: 2px;
+  }
+}
+
 .Progress {
   display: flex;
   align-items: center;
