+++ extension/ts/main-interface.js/components/Progress/demos/ControlledProgress.tsx
@@ -17,7 +17,7 @@ const ControlledProgress = (props: Props) => {
 
   return (
     <div
-      className={"Demo"}
+      className={"DemoControlledProgress"}
       style={{
         display: "flex",
         alignItems: "center",
