+++ extension/ts/main-interface.js/routes/Home.scss
@@ -1,0 +1,4 @@
+.Home {
+  height: 100%;
+  display: inline-block;
+}
