+++ extension/ts/main-interface.js/routes/Transition.tsx
@@ -1,12 +1,10 @@
 import * as React from "react";
 import { CSSTransition } from "react-transition-group";
 
-const Transition = () => {
+export const Transition = () => {
   return (
     <CSSTransition timeout={300}>
       <div></div>
     </CSSTransition>
   );
 };
-
-export default Transition;
